requirements.txt
gunicorn==20.1.0
PyExecJS==1.5.1
Flask==2.0.1
Flask-RESTful==0.3.9
requests==2.26.0